+++ core/src/test/java/org/realityforge/arez/test/ExternalApiTest.java
@@ -163,14 +163,14 @@ public class ExternalApiTest
     context.getSpy().addSpyEventHandler( handler );
 
     // Generate an event
-    context.createObservable( ValueUtil.randomString() );
+    context.createObservable();
 
     assertEquals( callCount.get(), 1 );
 
     context.getSpy().removeSpyEventHandler( handler );
 
     // Generate an event
-    context.createObservable( ValueUtil.randomString() );
+    context.createObservable();
 
     assertEquals( callCount.get(), 1 );
   }
@@ -181,7 +181,7 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     final AtomicInteger reactionCount = new AtomicInteger();
 
@@ -209,7 +209,7 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     final AtomicInteger reactionCount = new AtomicInteger();
 
@@ -238,10 +238,10 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable1 = context.createObservable( ValueUtil.randomString() );
-    final Observable observable2 = context.createObservable( ValueUtil.randomString() );
-    final Observable observable3 = context.createObservable( ValueUtil.randomString() );
-    final Observable observable4 = context.createObservable( ValueUtil.randomString() );
+    final Observable observable1 = context.createObservable();
+    final Observable observable2 = context.createObservable();
+    final Observable observable3 = context.createObservable();
+    final Observable observable4 = context.createObservable();
 
     final AtomicInteger reactionCount = new AtomicInteger();
 
@@ -299,7 +299,7 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     assertNotInTransaction( observable );
 
@@ -322,7 +322,7 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     assertNotInTransaction( observable );
 
@@ -344,7 +344,7 @@ public class ExternalApiTest
     throws Throwable
   {
     final ArezContext context = Arez.context();
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     assertNotInTransaction( observable );
 
@@ -373,7 +373,7 @@ public class ExternalApiTest
   {
     final ArezContext context = Arez.context();
 
-    final Observable observable = context.createObservable( ValueUtil.randomString() );
+    final Observable observable = context.createObservable();
 
     assertNotInTransaction( observable );
 
@@ -418,8 +418,8 @@ public class ExternalApiTest
 
     final ArezContext context1 = zone1.getContext();
     final ArezContext context2 = zone2.getContext();
-    final Observable observable1 = context1.createObservable( ValueUtil.randomString() );
-    final Observable observable2 = context2.createObservable( ValueUtil.randomString() );
+    final Observable observable1 = context1.createObservable();
+    final Observable observable2 = context2.createObservable();
 
     final AtomicInteger autorunCallCount1 = new AtomicInteger();
     final AtomicInteger autorunCallCount2 = new AtomicInteger();